pyproject.toml
[project]
name = "gitblog2"
version = "2.3.0"
description = "Git + Markdown = blog"
# authors = ["Henri Hannetel <henri.hannetel@pm.me>"]
# license = { file = "LICENSE" }
readme = "docs/index.md"
# documentation = "https://henritel.github.io/gitblog2"
# repository = "https://github.com/HenriTEL/gitblog2"
requires-python = ">=3.11"
dependencies = [
	"Jinja2 >= 3.1.2",
	"Markdown >= 3.4.1",
	"typer >=0.7.0",
	"feedgen >= 0.9.0",
	"requests >= 2.28.2",
	"gitpython >= 3.1.36",
]

[project.scripts]
gitblog2 = "gitblog2.cli:main"

[tool.uv.workspace]
# members = ["gitblog2"]
exclude = ["gitblog2/tests"]

[dependency-groups]
dev = [
  "pytest >=7.4.4",
  "ruff >=0.4.8",
  "mypy",
]

[tool.uv.sources]
mypy = { git = "https://github.com/python/mypy.git" }

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"

[tool.ruff.lint]
select = [
   "I",  # isort
   "F",  # Pyflakes
]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
enable_incomplete_feature = ["NewGenericSyntax"]
check_untyped_defs = true


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
